refactor(rimap-server): move resolve_download_dir_multi to boot::download_dir
