+++ crates/rimap-server/src/boot/mod.rs
@@ -3,5 +3,6 @@
 pub mod account_state;
 pub mod audit_init;
 pub mod discovery;
+pub(crate) mod download_dir;
 pub mod logging;
 pub mod registry;
